+++ filament-core/src/androidNativeMain/kotlin/dev/karmakrafts/filament/Thread.kt
@@ -14,9 +14,10 @@
  * limitations under the License.
  */
 
+@file:OptIn(ExperimentalForeignApi::class, UnsafeNumber::class)
+
 package dev.karmakrafts.filament
 
-import co.touchlab.stately.collections.ConcurrentMutableSet
 import kotlinx.cinterop.COpaquePointer
 import kotlinx.cinterop.ExperimentalForeignApi
 import kotlinx.cinterop.StableRef
@@ -28,7 +29,13 @@ import kotlinx.cinterop.memScoped
 import kotlinx.cinterop.ptr
 import kotlinx.cinterop.staticCFunction
 import kotlinx.cinterop.value
+import platform.posix.PTHREAD_CREATE_DETACHED
+import platform.posix.PTHREAD_CREATE_JOINABLE
 import platform.posix.nanosleep
+import platform.posix.pthread_attr_init
+import platform.posix.pthread_attr_setdetachstate
+import platform.posix.pthread_attr_setstacksize
+import platform.posix.pthread_attr_t
 import platform.posix.pthread_create
 import platform.posix.pthread_detach
 import platform.posix.pthread_gettid_np
@@ -38,59 +45,48 @@ import platform.posix.pthread_setname_np
 import platform.posix.pthread_t
 import platform.posix.pthread_tVar
 import platform.posix.timespec
+import kotlin.concurrent.atomics.AtomicBoolean
 import kotlin.native.concurrent.ThreadLocal
 
-@OptIn(UnsafeNumber::class)
-private val activeThreads: ConcurrentMutableSet<pthread_t> = ConcurrentMutableSet()
-
-@OptIn(UnsafeNumber::class)
-private val detachedThreads: ConcurrentMutableSet<pthread_t> = ConcurrentMutableSet()
-
-@PublishedApi
-internal actual val threadSupportsAffinity: Boolean = false
+@ThreadLocal
+internal var threadAffinity: Int = Thread.NO_AFFINITY
 
 @ThreadLocal
 internal var threadName: String? = null
 
-@OptIn(UnsafeNumber::class, ExperimentalForeignApi::class)
-private fun threadEntryPoint(userData: COpaquePointer?): COpaquePointer? {
-    userData?.asStableRef<() -> Unit>()?.apply {
-        get().invoke()
-        dispose()
+internal actual fun currentThread(): Thread = AndroidThread(pthread_self(), threadAffinity)
+
+private class AndroidThread( // @formatter:off
+    val handle: pthread_t,
+    override val affinity: Int
+) : Thread { // @formatter:on
+    private val _isAlive: AtomicBoolean = AtomicBoolean(true)
+    override val isAlive: Boolean
+        get() = _isAlive.load()
+
+    private val _isDetached: AtomicBoolean = AtomicBoolean(false)
+    override val isDetached: Boolean
+        get() = _isDetached.load()
+
+    override fun join() {
+        if (!_isAlive.compareAndSet(expectedValue = true, newValue = false)) return
+        pthread_join(handle, null)
     }
-    activeThreads -= pthread_self()
+
+    override fun detach() {
+        if (!_isDetached.compareAndSet(expectedValue = false, newValue = true)) return
+        pthread_detach(handle)
+    }
+}
+
+private fun threadTrampoline(userData: COpaquePointer?): COpaquePointer? {
+    val ref = userData?.asStableRef<() -> Unit>() ?: return null
+    ref.get()()
+    ref.dispose()
     return null
 }
 
-@OptIn(UnsafeNumber::class)
-internal actual fun currentThread(): ThreadHandle {
-    return NativeThreadHandle(pthread_self())
-}
-
-@OptIn(UnsafeNumber::class, ExperimentalForeignApi::class)
-internal actual fun createThread(function: () -> Unit): ThreadHandle = memScoped {
-    val handle = alloc<pthread_tVar>()
-    pthread_create(handle.ptr, null, staticCFunction(::threadEntryPoint), StableRef.create(function).asCPointer())
-    NativeThreadHandle(requireNotNull(handle.value) { "Could not create thread" }).apply {
-        activeThreads += value
-    }
-}
-
-@OptIn(UnsafeNumber::class, ExperimentalForeignApi::class)
-internal actual fun joinThread(handle: ThreadHandle) {
-    require(handle is NativeThreadHandle)
-    pthread_join(handle.value, null)
-}
-
-@OptIn(UnsafeNumber::class)
-internal actual fun detachThread(handle: ThreadHandle) {
-    require(handle is NativeThreadHandle)
-    pthread_detach(handle.value)
-    detachedThreads += handle.value
-}
-
 @PublishedApi
-@OptIn(UnsafeNumber::class)
 internal actual fun setThreadName(name: String?) {
     pthread_setname_np(pthread_self(), name)
     threadName = name
@@ -102,12 +98,10 @@ internal actual fun getThreadName(): String {
 }
 
 @PublishedApi
-@OptIn(UnsafeNumber::class)
 internal actual fun getThreadId(): ULong {
     return pthread_gettid_np(pthread_self()).toULong()
 }
 
-@OptIn(UnsafeNumber::class, ExperimentalForeignApi::class)
 internal actual fun sleepThread(millis: Long): Long = memScoped {
     val spec = alloc<timespec> {
         tv_sec = (millis / 1000).convert()
@@ -117,18 +111,31 @@ internal actual fun sleepThread(millis: Long): Long = memScoped {
     (spec.tv_sec * 1000).toLong() + (spec.tv_nsec / 1000000)
 }
 
-@OptIn(UnsafeNumber::class)
-internal actual fun isThreadAlive(handle: ThreadHandle): Boolean {
-    require(handle is NativeThreadHandle)
-    return handle.value in activeThreads
+actual fun Thread( // @formatter:off
+    affinity: Int,
+    stackSize: Long,
+    detached: Boolean,
+    function: () -> Unit
+): Thread = memScoped { // @formatter:on
+    val handle = alloc<pthread_tVar>()
+    val attributes = alloc<pthread_attr_t>()
+    check(pthread_attr_init(attributes.ptr) == 0) {
+        "Could not initialize thread attributes"
+    }
+    if (stackSize != Thread.DEFAULT_STACK_SIZE) {
+        check(pthread_attr_setstacksize(attributes.ptr, stackSize.convert()) == 0) {
+            "Could not set thread stack size"
+        }
+    }
+    val detachState = if (detached) PTHREAD_CREATE_DETACHED else PTHREAD_CREATE_JOINABLE
+    check(pthread_attr_setdetachstate(attributes.ptr, detachState) == 0) {
+        "Could not set thread detach state"
+    }
+    val trampolineAddress = staticCFunction(::threadTrampoline)
+    val closureAddress = StableRef.create(function).asCPointer()
+    check(pthread_create(handle.ptr, attributes.ptr, trampolineAddress, closureAddress) == 0) {
+        "Could not create thread"
+    }
+    threadAffinity = affinity
+    AndroidThread(handle.value, affinity)
 }
-
-@OptIn(UnsafeNumber::class)
-internal actual fun isThreadDetached(handle: ThreadHandle): Boolean {
-    require(handle is NativeThreadHandle)
-    return handle.value in detachedThreads
-}
-
-internal actual fun setThreadAffinity(logicalCore: Int) = Unit
-
-internal actual fun getThreadAffinity(): Int = 0