+++ filament-core/src/appleMain/kotlin/dev/karmakrafts/filament/Thread.kt
@@ -14,13 +14,13 @@
  * limitations under the License.
  */
 
+@file:OptIn(ExperimentalForeignApi::class, UnsafeNumber::class)
+
 package dev.karmakrafts.filament
 
-import co.touchlab.stately.collections.ConcurrentMutableSet
 import kotlinx.cinterop.ByteVar
 import kotlinx.cinterop.COpaquePointer
 import kotlinx.cinterop.ExperimentalForeignApi
-import kotlinx.cinterop.StableRef
 import kotlinx.cinterop.ULongVar
 import kotlinx.cinterop.UnsafeNumber
 import kotlinx.cinterop.alloc
@@ -31,82 +31,64 @@ import kotlinx.cinterop.memScoped
 import kotlinx.cinterop.pointed
 import kotlinx.cinterop.ptr
 import kotlinx.cinterop.reinterpret
-import kotlinx.cinterop.staticCFunction
 import kotlinx.cinterop.toKString
 import kotlinx.cinterop.value
-import platform.darwin.KERN_SUCCESS
-import platform.darwin.THREAD_AFFINITY_POLICY
-import platform.darwin.THREAD_AFFINITY_POLICY_COUNT
-import platform.darwin.thread_affinity_policy_data_t
-import platform.darwin.thread_policy_get
-import platform.darwin.thread_policy_set
 import platform.posix.nanosleep
-import platform.posix.pthread_create
 import platform.posix.pthread_detach
 import platform.posix.pthread_getname_np
 import platform.posix.pthread_join
-import platform.posix.pthread_mach_thread_np
 import platform.posix.pthread_self
 import platform.posix.pthread_setname_np
 import platform.posix.pthread_t
-import platform.posix.pthread_tVar
 import platform.posix.pthread_threadid_np
 import platform.posix.timespec
+import kotlin.concurrent.atomics.AtomicBoolean
+import kotlin.native.concurrent.ThreadLocal
 
-@OptIn(ExperimentalForeignApi::class)
-private val activeThreads: ConcurrentMutableSet<pthread_t> = ConcurrentMutableSet()
+@ThreadLocal
+internal var threadAffinity: Int = Thread.NO_AFFINITY
 
-@OptIn(ExperimentalForeignApi::class)
-private val detachedThreads: ConcurrentMutableSet<pthread_t> = ConcurrentMutableSet()
+internal actual fun currentThread(): Thread = AppleThread(checkNotNull(pthread_self()) {
+    "Could not get current thread handle"
+}, threadAffinity)
 
-@PublishedApi
-internal actual val threadSupportsAffinity: Boolean = false
+internal class AppleThread( // @formatter:off
+    val handle: pthread_t,
+    override val affinity: Int
+) : Thread { // @formatter:on
+    private val _isAlive: AtomicBoolean = AtomicBoolean(true)
+    override val isAlive: Boolean
+        get() = _isAlive.load()
 
-@OptIn(ExperimentalForeignApi::class)
-private fun threadEntryPoint(userData: COpaquePointer?): COpaquePointer? {
+    private val _isDetached: AtomicBoolean = AtomicBoolean(false)
+    override val isDetached: Boolean
+        get() = _isDetached.load()
+
+    override fun join() {
+        if (!_isAlive.compareAndSet(expectedValue = true, newValue = false)) return
+        pthread_join(handle, null)
+    }
+
+    override fun detach() {
+        if (!_isDetached.compareAndSet(expectedValue = false, newValue = true)) return
+        pthread_detach(handle)
+    }
+}
+
+internal fun threadTrampoline(userData: COpaquePointer?): COpaquePointer? {
     userData?.asStableRef<() -> Unit>()?.apply {
         get().invoke()
         dispose()
     }
-    activeThreads.remove(pthread_self())
     return null
 }
 
-@OptIn(ExperimentalForeignApi::class)
-internal actual fun currentThread(): ThreadHandle {
-    return NativeThreadHandle(requireNotNull(pthread_self()) { "Could not retrieve current thread" })
-}
-
-@OptIn(ExperimentalForeignApi::class)
-internal actual fun createThread(function: () -> Unit): ThreadHandle = memScoped {
-    val handle = alloc<pthread_tVar>()
-    pthread_create(handle.ptr, null, staticCFunction(::threadEntryPoint), StableRef.create(function).asCPointer())
-    NativeThreadHandle(requireNotNull(handle.value) { "Could not create thread" }).apply {
-        activeThreads.add(value)
-    }
-}
-
-@OptIn(ExperimentalForeignApi::class)
-internal actual fun joinThread(handle: ThreadHandle) {
-    require(handle is NativeThreadHandle)
-    pthread_join(handle.value, null)
-}
-
-@OptIn(ExperimentalForeignApi::class)
-internal actual fun detachThread(handle: ThreadHandle) {
-    require(handle is NativeThreadHandle)
-    pthread_detach(handle.value)
-    detachedThreads += handle.value
-}
-
 @PublishedApi
-@OptIn(ExperimentalForeignApi::class)
 internal actual fun setThreadName(name: String?) {
     pthread_setname_np(name)
 }
 
 @PublishedApi
-@OptIn(UnsafeNumber::class, ExperimentalForeignApi::class)
 internal actual fun getThreadName(): String = memScoped {
     val nameBuffer = allocArray<ByteVar>(4096).reinterpret<ByteVar>().pointed
     pthread_getname_np(pthread_self(), nameBuffer.ptr, 4096U)
@@ -114,62 +96,48 @@ internal actual fun getThreadName(): String = memScoped {
 }
 
 @PublishedApi
-@OptIn(ExperimentalForeignApi::class)
 internal actual fun getThreadId(): ULong = memScoped {
     val id = alloc<ULongVar>()
     pthread_threadid_np(pthread_self(), id.ptr)
     id.value
 }
 
-@OptIn(UnsafeNumber::class, ExperimentalForeignApi::class)
 internal actual fun sleepThread(millis: Long): Long = memScoped {
     val spec = alloc<timespec> {
         tv_sec = (millis / 1000).convert()
         tv_nsec = (millis % 1000000).convert()
     }
     nanosleep(spec.ptr, spec.ptr)
-    (spec.tv_sec * 1000).toLong() + (spec.tv_nsec / 1000000)
-}
-
-@OptIn(ExperimentalForeignApi::class)
-internal actual fun isThreadAlive(handle: ThreadHandle): Boolean {
-    require(handle is NativeThreadHandle)
-    return handle.value in activeThreads
-}
-
-@OptIn(ExperimentalForeignApi::class)
-internal actual fun isThreadDetached(handle: ThreadHandle): Boolean {
-    require(handle is NativeThreadHandle)
-    return handle.value in detachedThreads
+    (spec.tv_sec * 1000).convert<Long>() + (spec.tv_nsec / 1000000)
 }
 
 // The Darwin kernel does not support per-core affinity like Linux, so we emulate
 // core mappings through thread affinity tags
-@OptIn(ExperimentalForeignApi::class)
-internal actual fun setThreadAffinity(logicalCore: Int) = memScoped {
-    val handle = pthread_mach_thread_np(pthread_self())
-    val policy = alloc<thread_affinity_policy_data_t> {
-        affinity_tag = logicalCore
-    }
-    val result = thread_policy_set( // @formatter:off
-        thread = handle,
-        flavor = THREAD_AFFINITY_POLICY.toUInt(),
-        policy_info = policy.ptr.reinterpret(),
-        policy_infoCnt = THREAD_AFFINITY_POLICY_COUNT
-    ) // @formatter:on
-    check(result == KERN_SUCCESS) { "Could not update thread affinity tag" }
-}
-
-@OptIn(ExperimentalForeignApi::class)
-internal actual fun getThreadAffinity(): Int = memScoped {
-    val handle = pthread_mach_thread_np(pthread_self())
-    val policy = alloc<thread_affinity_policy_data_t>()
-    thread_policy_get(
-        thread = handle,
-        flavor = THREAD_AFFINITY_POLICY.toUInt(),
-        policy_info = policy.ptr.reinterpret(),
-        policy_infoCnt = null,
-        get_default = null
-    )
-    return policy.affinity_tag
-}
+//@OptIn(ExperimentalForeignApi::class)
+//internal actual fun setThreadAffinity(logicalCore: Int) = memScoped {
+//    val handle = pthread_mach_thread_np(pthread_self())
+//    val policy = alloc<thread_affinity_policy_data_t> {
+//        affinity_tag = logicalCore
+//    }
+//    val result = thread_policy_set( // @formatter:off
+//        thread = handle,
+//        flavor = THREAD_AFFINITY_POLICY.toUInt(),
+//        policy_info = policy.ptr.reinterpret(),
+//        policy_infoCnt = THREAD_AFFINITY_POLICY_COUNT
+//    ) // @formatter:on
+//    check(result == KERN_SUCCESS) { "Could not update thread affinity tag" }
+//}
+//
+//@OptIn(ExperimentalForeignApi::class)
+//internal actual fun getThreadAffinity(): Int = memScoped {
+//    val handle = pthread_mach_thread_np(pthread_self())
+//    val policy = alloc<thread_affinity_policy_data_t>()
+//    thread_policy_get(
+//        thread = handle,
+//        flavor = THREAD_AFFINITY_POLICY.toUInt(),
+//        policy_info = policy.ptr.reinterpret(),
+//        policy_infoCnt = null,
+//        get_default = null
+//    )
+//    return policy.affinity_tag
+//}