+++ filament-core/src/nativeMain/kotlin/dev/karmakrafts/filament/Thread.kt
@@ -17,13 +17,12 @@
 package dev.karmakrafts.filament
 
 import kotlinx.cinterop.ExperimentalForeignApi
-import platform.posix.pthread_t
 import platform.posix.sched_yield
 
-@OptIn(ExperimentalForeignApi::class)
-internal data class NativeThreadHandle(
-    val value: pthread_t
-) : ThreadHandle
+//@OptIn(ExperimentalForeignApi::class)
+//internal data class NativeThreadHandle(
+//    val value: pthread_t
+//) : ThreadHandle
 
 @OptIn(ExperimentalForeignApi::class)
 internal actual fun yieldThread() {